renaming to v1 and ClusterStatus
