+++ api/v1/cluster_test.go
@@ -1,4 +1,4 @@
-package cluster
+package v1
 
 import (
 	"time"
