+++ api/v1/clusterdata_test.go
@@ -1,4 +1,4 @@
-package cluster
+package v1
 
 import (
 	. "github.com/onsi/ginkgo/v2"
