+++ api/v1/duration.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cluster
+package v1
 
 import (
 	"encoding/json"
