+++ api/v1/random_utils_test.go
@@ -1,4 +1,4 @@
-package cluster
+package v1
 
 import (
 	"fmt"
